Add a host-only build module for guava-testlib
Guava testlib is a set of Java classes for more convenient unit testing.
Many of the utilities in this library eliminate the need for writing
trivial tests such as null pointer, equality, and class sanity tests.

Note that this change only adds a module for use in host-side tests to
avoid maintaining a device target.

Test: m guava-testlib
Test: Used library and ran unit tests
Bug: 148404241
Change-Id: Ifa3e5a7818abab25a5f1b5aca845daf8caddb600