`guava` target is alias depending on configuration
Instead of requiring modules to explicitly depend on one of the guava
targets, we provide an alias that matches the correct dependency
depending on what Bazel configuration is available for that dependency
edge.

In particular, this allows error_prone to build as a host module and
depend on the host-only Guava library.

Bug: 215229742
Test: build/bazel/bp2build.sh
Change-Id: I714834695a124316c1d80cd734555d177f608bd6